+++ examples/platforms/cc2538/radio.c
@@ -36,7 +36,6 @@
 #include <openthread/openthread.h>
 #include <openthread/platform/alarm-milli.h>
 #include <openthread/platform/diag.h>
-#include <openthread/platform/platform.h>
 #include <openthread/platform/radio.h>
 
 #include "platform-cc2538.h"
