+++ examples/platforms/cc2652/alarm.c
@@ -42,8 +42,6 @@
 
 #include <openthread/platform/alarm-milli.h>
 #include <openthread/platform/diag.h>
-#include <openthread/platform/platform.h>
-
 
 #include "platform-cc2652.h"
