+++ examples/platforms/gp712/alarm.c
@@ -37,7 +37,6 @@
 #include <sys/time.h>
 
 #include <openthread/openthread.h>
-#include <openthread/platform/platform.h>
 #include <openthread/platform/alarm-milli.h>
 #include <openthread/platform/diag.h>
 #include "alarm_qorvo.h"
