+++ examples/platforms/kw41z/alarm.c
@@ -38,7 +38,6 @@
 #include "fsl_pit.h"
 
 #include "openthread/openthread.h"
-#include "openthread/platform/platform.h"
 #include "openthread/platform/alarm-milli.h"
 #include "openthread/platform/diag.h"
