+++ examples/platforms/platform.h
